refactor: Change module structure
